fix: downgrading image-webpack-loader to fix linux arch bug (#147)
There’s a documented issue in “imagemin-gifsicle” where a recent update (5.0.0) removed the binaries for linux distributions:

https://github.com/imagemin/gifsicle-bin/issues/113

This causes it to explode in CI via our direct dependency on image-webpack-loader.  The 7.0.0 version of image-webpack-loader relies on version 5 of gifsicle.  If we downgrade back to 6.0.0, we get gifsicle 4.0.1 instead, which still has the binaries.