feat: upgrade build ecosystem and webpack to v5 (#177)
* feat(webpack): upgrade to webpack 5 (#128)

BREAKING CHANGE: upgrades @edx/frontend-build to use Webpack 5, which includes new features such as module federation.

* fix: add full relative route

* fix: update react and react-dom in package-lock.json

* fix: publish alpha

* fix: npm and github secrets

* fix: bumping webpack to 5.21.2

* fix: bumping html-webpack-plugin, webpack-cli

* fix: downgrading webpack-cli to ^3

See issues in this thread: https://github.com/webpack/webpack-dev-server/issues/2029

Solution for now seems to be to downgrade from 4 to 3 to get the cli to work with webpack 5.

* fix: bumping dotenv-webpack

Version 5 supports webpack 5.

* test: testing process.env variable usage

* fix: bumping to webpack-dev-server 4.0.0 beta

The 3.x version of the dev server is not compatible with webpack 5 or the latest webpack-cli.  Everything explodes.  Also bumping react-dev-utils to get the latest verison of the webpackHotDevClient.

* build: fix broken devServer config

Put many of the options in the wrong place.  The “inline” option is also deprecated and needed to be removed.

* test: adding RTL-sensitive CSS so we can test it

* fix: replacing postcss-rtl with postcss-rtlcss

Also upgrading other webpack dependencies.

* fix: bumping resolve-url-loader to v4.x

Hoping this fixes some of the postcss version skew issues.

* fix: updating webpack.dev-stage.config.js to match the non-stage version

* fix: bumping resolve-url-loader to v5 beta

This bumps its version of postcss to 8, which I’m hoping will solve dependency issues where dependent MFEs are getting postcss 7 because they don’t explictly add postcss as a dependency, which is something I’m trying to avoid.

* test: fixing a broken test

Not sure how the test variable/non-existent variable code wasn’t already in the snapshot - all this PR added was the right-aligned text.

* fix: reconciling package-lock.json with rebased changes

* test: adding TEST_VARIABLE to test env and snapshot

* build: re-generating package-lock.json after rebasing to keep it clean

* build: removing unncessary release.yml config

The ‘branches’ section here is actually a subset of the default, and not really meaningful for us to have added.

* fix: trivial fix to see if alpha release version will bump

* fix: another trivial change to test semantic-release

Co-authored-by: Adam Stankiewicz <astankiewicz@edx.org>
Co-authored-by: Adam Stankiewicz <agstanki@gmail.com>