feat: use new jsx transform
This updates the @babel/preset-react configuration to use the new JSX transform, which no longer requires React to be imported into source files.

It also updates our eslint-config to no longer require that React be imported into source files as well.

This works because our babel (7.9.0+) and react (16.14.0+) versions are high enough that they can take advantage of the new JSX transform.

Note that this means this version of frontend-build has a new peer dependency on react > 16.14.0, to ensure that it has the new JSX transform available.