feat: JavaScript-based environment configuration (#200)
This adds a fairly simple mechanism to the webpack config that resolves a package named env.config to a local env.config.js file in the app. If such a file doesn't exit, it falls back to an empty env.config.js file in frontend-build so that webpack doesn't break.

This is the groundwork required for us to use JS-based environment configuration. This would have numerous benefits, most importantly, letting us use non-string values in our config (like booleans and numbers) as well as importing other resources to be used as config, such as alternate implementations of services in frontend-platform.

Note: The code in this PR makes it technically possible for us to optionally import such a file, but the mechanism isn't really useful yet because it requires support in both frontend-platform and our build processes. Stay tuned.

Commits:

* feat: adding the basic mechanism for using a JS file for env config

- Add an alias and fallback for "env.config" in the webpack.common.config.js file.
- Put an empty fallback file in the config directory in case an MFE doesn't specify its own.

The idea is that an MFE (or probably frontend-platform) can provide an env.config.js file which will be linked in if it exists.  This is an alternative to .env files.

* test: use a sample env.config.js file in the example app

This commit also organizes the tests in the example app a bit.

* fix: fixing eslint mis-configuration

Also cleaning up example app linting errors.

* docs: adding an ADR describing the decision to move to JS env config

* test: fixing broken tests and jest config

The jest config needs to have the same sort of fallback behavior that webpack uses.  It doesn't use webpack itself, just babel.

* refactor: use false value for fallback instead of a file

If the "fallback" webpack resolve parameter is set to `false`, then it resolves the module with an empty object.  This is perfect for us.

It also lets us move our fallback file into the jest subdirectory, since we still need a physical file when jest gets run.

* docs: mention tubular in ADR

* fix: removing unnecessary .babelrc.js file

We don't actually need this file in order to run the build; it pulls it from the config directory on its own.

* test: fixing SASS variable name in example styles

Making the variable name match the element it's being applied to.