feat: Include OPTIMIZELY_PROJECT_ID in the HtmlWebpackPlugic dev config.
This will allow dev environments to include the Optimizely script snippet via index.html,
which we know we need in at least two consuming MFEs.
This should be fine for any downstream consumers that don't use Optimizely, because
we tend to conditionally include an Optimizely script in index.html based on the existence of
an OPTIMIZELY_PROJECT_ID variable in the environment (i.e. in `.env.development`). So if
that variable is not present by default (which it shouldn't be), no Optimizely script is included.