fix: Exclude openedx scoped packages.
We are moving packages to the `openedx` scope but those packages will
also not be transpiled and we'll need to handle them in the same way.

Marking this as a fix because we want to release a new version with this
exclusion before we start moving more packages to the new scope.