feat: get rid of bundle analyzer on by default in prod
Every run of `npm run build` generates a report.html file that includes
file size stats. It's useless in most cases. We suggest to remove it
(along with the dependency) in all MFEs.

This shaves 14s (out of 174s) of the build on my laptop for the learning
MFE build.