chore: lighter prod dependencies
We move development dependencies to devDependencies, such that `npm
install --omit=dev` becomes faster and more lightweight. In the case of
frontend-app-profile, this changes shaves off a few hundred MB of disk
space.

Note the move of @babel/preset-typescript from devDependencies to
dependencies: this fixes MFE build with yarn, which was previously
failing:

    git clone https://github.com/openedx/frontend-app-profile/
    cd frontend-app-profile
    yarn install
    npm run build

Was resulting in the following error:

	ERROR in ./src/index.jsx
	Module build failed (from ./node_modules/babel-loader/lib/index.js):
	Error: Cannot find package '@babel/preset-typescript' imported from node_modules/@openedx/frontend-build/config/babel-virtual-resolve-base.js

This is because yarn does not install devDependencies of package
dependencies, as opposed to npm.